+++ src/flexi/flexi_class.c
@@ -16,10 +16,7 @@ static int _create_class_record(struct flexi_Context_t *pCtx, const char *zClass
     int result;
     if (!pCtx->pStmts[STMT_INS_CLS])
     {
-        CHECK_CALL(sqlite3_prepare_v2(pCtx->db,
-                                      "insert into [.classes] (NameID) values (:1);",
-                                      -1, &pCtx->pStmts[STMT_INS_CLS],
-                                      NULL));
+        CHECK_STMT_PREPARE(pCtx->db, "insert into [.classes] (NameID) values (:1);", &pCtx->pStmts[STMT_INS_CLS]);
     }
     CHECK_CALL(sqlite3_reset(pCtx->pStmts[STMT_INS_CLS]));
     sqlite3_int64 lClassNameID;
@@ -51,11 +48,12 @@ static int _parseSpecialProperties(struct flexi_ClassDef_t *pClassDef, const cha
     };
     char *zSql = NULL;
     zSql = sqlite3_mprintf("select ");
-    const char *zp = zSpecProps[0];
+
     char sep[] = " ";
-    for (int ii = 0; ii < ARRAY_LEN(zSpecProps); ii++, zp++)
+    for (int ii = 0; ii < ARRAY_LEN(zSpecProps); ii++)
     {
         char *zTemp = zSql;
+        const char *zp = zSpecProps[ii];
         zSql = sqlite3_mprintf("%s%s\n json_extract(:1, '$.specialProperties.%s.$id') as %s_id,"
                                        "json_extract(:1, '$.specialProperties.%s.$name') as %s_name",
                                zTemp, sep, zp, zp, zp, zp);
@@ -63,7 +61,7 @@ static int _parseSpecialProperties(struct flexi_ClassDef_t *pClassDef, const cha
         sqlite3_free(zTemp);
     }
 
-    CHECK_CALL(sqlite3_prepare_v2(pClassDef->pCtx->db, zSql, -1, &pStmt, NULL));
+    CHECK_STMT_PREPARE(pClassDef->pCtx->db, zSql, &pStmt);
     CHECK_CALL(sqlite3_bind_text(pStmt, 1, zClassDefJson, -1, NULL));
     CHECK_STMT(sqlite3_step(pStmt));
     if (result == SQLITE_ROW)
@@ -107,7 +105,7 @@ static int _parseRangeProperties(struct flexi_ClassDef_t *pClassDef, const char 
         sep = ',';
     }
 
-    CHECK_CALL(sqlite3_prepare_v2(pClassDef->pCtx->db, zSql, -1, &pStmt, NULL));
+    CHECK_STMT_PREPARE(pClassDef->pCtx->db, zSql, &pStmt);
     CHECK_CALL(sqlite3_bind_text(pStmt, 1, zClassDefJson, -1, NULL));
     CHECK_STMT(sqlite3_step(pStmt));
     if (result == SQLITE_ROW)
@@ -151,7 +149,7 @@ static int _parseFullTextProperties(struct flexi_ClassDef_t *pClassDef, const ch
     }
 
     sqlite3_stmt *pStmt = NULL;
-    CHECK_CALL(sqlite3_prepare_v2(pClassDef->pCtx->db, zSql, -1, &pStmt, NULL));
+    CHECK_STMT_PREPARE(pClassDef->pCtx->db, zSql, &pStmt);
     CHECK_CALL(sqlite3_bind_text(pStmt, 1, zClassDefJson, -1, NULL));
     CHECK_STMT(sqlite3_step(pStmt));
     if (result == SQLITE_ROW)
@@ -190,7 +188,7 @@ static int _parseMixins(struct flexi_ClassDef_t *pClassDef, const char *zClassDe
             "json_extract(value, '$.dynamic.selectorProp.$name'), " // 3
             "json_extract(value, '$.dynamic.rules') " // 4
             "from json_each(:1, '$.mixins')";
-    CHECK_CALL(sqlite3_prepare_v2(pClassDef->pCtx->db, zSql, -1, &pStmt, NULL));
+    CHECK_STMT_PREPARE(pClassDef->pCtx->db, zSql, &pStmt);
     CHECK_CALL(sqlite3_bind_text(pStmt, 1, zClassDefJson, -1, NULL));
     while (true)
     {
@@ -223,7 +221,7 @@ static int _parseMixins(struct flexi_ClassDef_t *pClassDef, const char *zClassDe
                 "json_extract(value, '$.classRef.$id') as classId," // 1
                 "json_extract(value, '$.classRef.$name') as className" // 2
                 "from json_each(:1);";
-        CHECK_CALL(sqlite3_prepare_v2(pClassDef->pCtx->db, zRulesSql, -1, &pRulesStmt, NULL));
+        CHECK_STMT_PREPARE(pClassDef->pCtx->db, zRulesSql, &pRulesStmt);
         CHECK_CALL(sqlite3_bind_text(pRulesStmt, 1, zRulesJson, -1, NULL));
         while (true)
         {
@@ -404,213 +402,6 @@ int flexi_class_create(struct flexi_Context_t *pCtx, const char *zClassName,
 
     CHECK_CALL(_flexi_ClassDef_applyNewDef(pCtx, lClassID, zClassDef, bCreateVTable, INVALID_DATA_ABORT, pzError));
 
-    // TODO Remove
-
-    //    char *sbClassDefJSON = sqlite3_mprintf("{\"properties\":{");
-
-    //    sqlite3_int64 lClassNameID;
-    //    CHECK_CALL(flexi_Context_insertName(pCtx, zClassName, &lClassNameID));
-
-    // insert into .classes
-    //    {
-    //        const char *zInsClsSQL = "insert into [.classes] (NameID) values (:1);";
-    //
-    //        CHECK_CALL(sqlite3_prepare_v2(pCtx->db, zInsClsSQL, -1, &pInsClsStmt, NULL));
-    //        sqlite3_bind_int64(pInsClsStmt, 1, lClassNameID);
-    //        int stepResult = sqlite3_step(pInsClsStmt);
-    //        if (stepResult != SQLITE_DONE)
-    //        {
-    //            result = stepResult;
-    //            goto ONERROR;
-    //        }
-    //    }
-
-    //    sqlite3_int64 iClassID;
-    //    {
-    //        if (pCtx->pStmts[STMT_SEL_CLS_BY_NAME] == NULL)
-    //        {
-    //            CHECK_CALL(sqlite3_prepare_v2(
-    //                    pCtx->db,
-    //                    "select ClassID from [.classes] where NameID = (select NameID from [.names] where [Value] = :1 limit 1);",
-    //                    -1, &pCtx->pStmts[STMT_SEL_CLS_BY_NAME], NULL));
-    //        }
-    //
-    //        sqlite3_stmt *p = pCtx->pStmts[STMT_SEL_CLS_BY_NAME];
-    //        assert(p);
-    //        sqlite3_reset(p);
-    //        sqlite3_bind_text(p, 1, zClassName, -1, NULL);
-    //        int stepRes = sqlite3_step(p);
-    //        if (stepRes != SQLITE_ROW)
-    //        {
-    //            result = stepRes;
-    //            goto ONERROR;
-    //        }
-    //
-    //        iClassID = sqlite3_column_int64(p, 0);
-    //    }
-
-    //    int xCtloMask = 0;
-    //
-    //    const char *zInsPropSQL = "insert into [flexi_prop] (NameID, ClassID, ctlv, ctlvPlan)"
-    //            " values (:1, :2, :3, :4);";
-    //    CHECK_CALL(sqlite3_prepare_v2(pCtx->db, zInsPropSQL, -1, &pInsPropStmt, NULL));
-    //
-    //    // Prepare JSON processing
-    //    const char *zExtractPropSQL = "select "
-    //            "coalesce(json_extract(value, '$.index'), 'none') as indexed," // 0
-    //            //    subType // 1
-    //            //    minOccurences // 2
-    //            //    maxOccurences // 3
-    //            "coalesce(json_extract(value, '$.rules.type'), 'text') as type," // 4
-    //            "key as prop_name," // 5
-    //            "value as prop_def," // 6 - Original property definition JSON
-    //            "coalesce(json_extract(value, '$.noTrackChanges'), 0) as indexed," // 7
-    //            //    enumDef
-    //            //    refDef
-    //            //    $renameTo
-    //            //    $drop
-    //            //    rules.maxLength
-    //            //    rules.minValue
-    //            //    rules.maxValue
-    //            //    rules.regex
-    //            " from json_each(:1, '$.properties');";
-    //
-    //    // Need to remove leading and trailing quotes
-    //    int iJSONLen = (int) strlen(zClassDef);
-    //    CHECK_CALL(sqlite3_prepare_v2(pCtx->db, zExtractPropSQL, -1, &pExtractProps, NULL));
-    //    CHECK_CALL(sqlite3_bind_text(pExtractProps, 1, zClassDef + sizeof(char), iJSONLen - 2, NULL));
-    //
-    //    int iPropCnt = 0;
-    //
-    //    // Load property definitions from JSON
-    //    while (1)
-    //    {
-    //        int iStep = sqlite3_step(pExtractProps);
-    //        if (iStep == SQLITE_DONE)
-    //            break;
-    //
-    //        if (iStep != SQLITE_ROW)
-    //        {
-    //            result = iStep;
-    //            goto ONERROR;
-    //        }
-    //
-    //        memset(&dProp, 0, sizeof(dProp));
-    //        dProp.bIndexed = (char) sqlite3_column_int(pExtractProps, 0);
-    //        dProp.bUnique = (char) sqlite3_column_int(pExtractProps, 1);
-    //        dProp.bFullTextIndex = (char) sqlite3_column_int(pExtractProps, 2);
-    //        dProp.xRole = (short int) sqlite3_column_int(pExtractProps, 3);
-    //        dProp.type = (short int) sqlite3_column_int(pExtractProps, 4);
-    //
-    //        sqlite3_free((void *) zPropDefJSON);
-    //
-    //        CHECK_CALL(getColumnAsText(&dProp.name.name, pExtractProps, 5));
-    //        CHECK_CALL(getColumnAsText(&zPropDefJSON, pExtractProps, 6));
-    //
-    //        // Property control flags which regulate actual indexing and other settings
-    //        int xCtlv = 0;
-    //
-    //        // Planned (postponed for future) property control flags which will be applied later
-    //        // when enough statistics accumulated about best index strategy.
-    //        // Typically, this will happen when database size reaches few megabytes and 1K-5K records
-    //        // On smaller databases there is no real point to apply indexing to the full extent
-    //        // Plus, in the database schema lifetime initial period is usually associated with heavy refactoring
-    //        // and data restructuring.
-    //        // Taking into account these 2 considerations, we will remember user settings for desired indexing
-    //        // (in ctlvPlan) but currently apply only settings for unique values (as it is mostly constraint, rather
-    //        // than indexing)
-    //        int xCtlvPlan = 0;
-    //
-    //        switch (dProp.type)
-    //        {
-    //            // These property types can be searched by range, can be indexed and can be unique
-    //            case PROP_TYPE_DECIMAL:
-    //            case PROP_TYPE_NUMBER:
-    //            case PROP_TYPE_DATETIME:
-    //            case PROP_TYPE_INTEGER:
-    //
-    //                // These property types can be indexed
-    //            case PROP_TYPE_BINARY:
-    //            case PROP_TYPE_NAME:
-    //            case PROP_TYPE_ENUM:
-    //            case PROP_TYPE_UUID:
-    //                //                if (dProp.bUnique || (dProp.xRole & PROP_ROLE_ID) || (dProp.xRole & PROP_ROLE_NAME))
-    //                //                {
-    //                //                    xCtlv |= CTLV_UNIQUE_INDEX;
-    //                //                    xCtlvPlan |= CTLV_UNIQUE_INDEX;
-    //                //                }
-    //                // Note: no break here;
-    //
-    //            case PROP_TYPE_TEXT:
-    //                if (dProp.bIndexed && dProp.maxLength <= 30)
-    //                    xCtlvPlan |= CTLV_INDEX;
-    //                if (dProp.bFullTextIndex)
-    //                    xCtlvPlan |= CTLV_FULL_TEXT_INDEX;
-    //
-    //                break;
-    //
-    //            default:
-    //                break;
-    //        }
-    //
-    //        sqlite3_int64 lPropNameID;
-    //        CHECK_CALL(flexi_Context_insertName(pCtx, dProp.name.name, &lPropNameID));
-    //
-    //        {
-    //            sqlite3_reset(pInsPropStmt);
-    //            sqlite3_bind_int64(pInsPropStmt, 1, lPropNameID);
-    //            sqlite3_bind_int64(pInsPropStmt, 2, iClassID);
-    //            sqlite3_bind_int(pInsPropStmt, 3, xCtlv);
-    //            sqlite3_bind_int(pInsPropStmt, 4, xCtlvPlan);
-    //            int stepResult = sqlite3_step(pInsPropStmt);
-    //            if (stepResult != SQLITE_DONE)
-    //            {
-    //                result = stepResult;
-    //                goto ONERROR;
-    //            }
-    //        }
-    //
-    //        // Get new property ID
-    //        sqlite3_int64 iPropID;
-    //        CHECK_CALL(flexi_Context_getPropIdByClassAndNameIds(pCtx, iClassID, lPropNameID, &iPropID));
-    //        if (iPropCnt != 0)
-    //        {
-    //            void *pTmp = sbClassDefJSON;
-    //            sbClassDefJSON = sqlite3_mprintf("%s,", pTmp);
-    //            sqlite3_free(pTmp);
-    //        }
-    //
-    //        {
-    //            void *pTmp = sbClassDefJSON;
-    //            sbClassDefJSON = sqlite3_mprintf("%s\"%lld\":%s", pTmp, iPropID, zPropDefJSON);
-    //            sqlite3_free(pTmp);
-    //        }
-    //
-    //        iPropCnt++;
-    //    }
-    //
-    //    {
-    //        void *pTmp = sbClassDefJSON;
-    //        sbClassDefJSON = sqlite3_mprintf("%s}}", pTmp);
-    //        sqlite3_free(pTmp);
-    //    }
-    //
-    //    // Update class with new JSON data
-    //    const char *zUpdClsSQL = "update [.classes] set Data = :1, ctloMask= :2 where ClassID = :3";
-    //    CHECK_CALL(sqlite3_prepare_v2(pCtx->db, zUpdClsSQL, -1, &pUpdClsStmt, NULL));
-    //    sqlite3_bind_text(pUpdClsStmt, 1, sbClassDefJSON, (int) strlen(sbClassDefJSON), NULL);
-    //    sqlite3_bind_int(pUpdClsStmt, 2, xCtloMask);
-    //    sqlite3_bind_int64(pUpdClsStmt, 3, iClassID);
-    //    int updResult = sqlite3_step(pUpdClsStmt);
-    //    if (updResult != SQLITE_DONE)
-    //    {
-    //        result = updResult;
-    //        goto ONERROR;
-    //    }
-
-    // TODO
-    //    CHECK_CALL(flexi_class_def_load(db, pAux, zClassName, ppVTab, pzErr));
-
     result = SQLITE_OK;
 
     goto EXIT;
@@ -714,7 +505,7 @@ int flexi_class_alter_func(
 {
     assert(argc >= 2 && argc <= 4);
 
-    int result;
+    int result = SQLITE_OK;
     const char *zError = NULL;
 
     // 1st arg: class name
@@ -870,8 +661,8 @@ int flexi_class_rename(struct flexi_Context_t *pCtx, sqlite3_int64 iOldClassID, 
     {
         const char *zErrMsg;
         sqlite3_stmt *pStmt;
-        CHECK_CALL(sqlite3_prepare_v2(pCtx->db, "update [.classes] set NameID = :1 "
-                "where ClassID = :2;", -1, &pCtx->pStmts[STMT_CLS_RENAME], NULL));
+        CHECK_STMT_PREPARE(pCtx->db, "update [.classes] set NameID = :1 "
+                "where ClassID = :2;", &pCtx->pStmts[STMT_CLS_RENAME]);
     }
     CHECK_CALL(sqlite3_reset(pCtx->pStmts[STMT_CLS_RENAME]));
     sqlite3_bind_int64(pCtx->pStmts[STMT_CLS_RENAME], 1, lNewNameID);
@@ -1077,7 +868,7 @@ int flexi_class_def_parse(struct flexi_ClassDef_t *pClassDef,
 
     // Load properties
     char *zPropSql = "select key as Name, value as Definition from json_each(:1, '$.properties');";
-    CHECK_CALL(sqlite3_prepare_v2(pClassDef->pCtx->db, zPropSql, -1, &pStmt, NULL));
+    CHECK_STMT_PREPARE(pClassDef->pCtx->db, zPropSql,  &pStmt);
     CHECK_CALL(sqlite3_bind_text(pStmt, 1, zClassDefJson, -1, NULL));
     CHECK_CALL(_parseProperties(pClassDef, pStmt, 0, 1, -1, -1, -1));
 
@@ -1125,7 +916,7 @@ int flexi_class_def_load(struct flexi_Context_t *pCtx, sqlite3_int64 lClassID, s
             "Data as Definition " // 5
             "from [.classes] "
             "where ClassID = :1;";
-    CHECK_CALL(sqlite3_prepare_v2(pCtx->db, zGetClassSQL, -1, &pGetClassStmt, NULL));
+    CHECK_STMT_PREPARE(pCtx->db, zGetClassSQL,  &pGetClassStmt);
     sqlite3_bind_int64(pGetClassStmt, 1, lClassID);
     result = sqlite3_step(pGetClassStmt);
     if (result == SQLITE_DONE)
@@ -1150,7 +941,7 @@ int flexi_class_def_load(struct flexi_Context_t *pCtx, sqlite3_int64 lClassID, s
     // Load properties from flexi_prop
     if (!pCtx->pStmts[STMT_LOAD_CLS_PROP])
     {
-        CHECK_CALL(sqlite3_prepare_v2(pCtx->db, "select "
+        CHECK_STMT_PREPARE(pCtx->db, "select "
                                               "PropertyID," // 0
                                               "Class, " // 1
                                               "NameID, " // 2
@@ -1158,9 +949,8 @@ int flexi_class_def_load(struct flexi_Context_t *pCtx, sqlite3_int64 lClassID, s
                                               "ctlv," // 4
                                               "ctlvPlan," // 5
                                               "Definition" // 6
-                                              " from [flexi_prop] where ClassID=:1", -1,
-                                      &pCtx->pStmts[STMT_LOAD_CLS_PROP], NULL));
-
+                                              " from [flexi_prop] where ClassID=:1",
+                           &pCtx->pStmts[STMT_LOAD_CLS_PROP]);
     }
     CHECK_CALL(sqlite3_reset(pCtx->pStmts[STMT_LOAD_CLS_PROP]));
     CHECK_CALL(sqlite3_bind_int64(pCtx->pStmts[STMT_LOAD_CLS_PROP], 1, lClassID));
@@ -1188,9 +978,9 @@ int flexi_class_def_load(struct flexi_Context_t *pCtx, sqlite3_int64 lClassID, s
     return result;
 }
 
-void flexi_schema_func(sqlite3_context *context,
-                       int argc,
-                       sqlite3_value **argv)
+int flexi_schema_func(sqlite3_context *context,
+                      int argc,
+                      sqlite3_value **argv)
 {
     int result;
 
@@ -1212,7 +1002,7 @@ void flexi_schema_func(sqlite3_context *context,
         if (argc == 2)
             bCreateVTable = sqlite3_value_int(argv[1]) != 0;
         void *pCtx = sqlite3_user_data(context);
-        CHECK_CALL(sqlite3_prepare_v2(db, "select value, key from json_each(:1)", -1, &pStmt, NULL));
+        CHECK_STMT_PREPARE(db, "select value, key from json_each(:1)", &pStmt);
         CHECK_CALL(sqlite3_bind_value(pStmt, 1, sqlite3_value_dup(argv[0])));
 
         while ((result = sqlite3_step(pStmt)) == SQLITE_ROW)
@@ -1246,6 +1036,6 @@ void flexi_schema_func(sqlite3_context *context,
     EXIT:
     sqlite3_finalize(pStmt);
     sqlite3_free(zClassDef);
-    sqlite3_free(zClassName);
+    return result;
 }